add claim gas estimate cli
