+++ packages/cardpay-cli/rewards/register-rewardee-gas-estimate.ts
@@ -2,7 +2,7 @@ import { Argv } from 'yargs';
 import { getWeb3, NETWORK_OPTION_LAYER_2, getWeb3Opts, FROM_OPTION } from '../utils';
 import { Arguments, CommandModule } from 'yargs';
 import { getSDK } from '@cardstack/cardpay-sdk';
-import { fromWei} from 'web3-utils';
+import { fromWei } from 'web3-utils';
 
 export default {
   command: 'register-rewardee-gas-estimate <prepaidCard> <rewardProgramId>',
@@ -28,14 +28,10 @@ export default {
     };
     let web3 = await getWeb3(network, getWeb3Opts(args));
     let rewardManager = await getSDK('RewardManager', web3);
-    let estimate = await rewardManager.registerRewardeeGasEstimate(
-        prepaidCard, rewardProgramId
-    )
+    let estimate = await rewardManager.registerRewardeeGasEstimate(prepaidCard, rewardProgramId);
     console.log(
       `The gas estimate for registering a rewardee for ${rewardProgramId} is
-      ${fromWei(
-        estimate
-      )}`
+      ${fromWei(estimate)}`
     );
   },
 } as CommandModule;
